chore(tests): meaningful split between test.js and playwright.spec.js
CLI handling goes to test.js.
Everything about running tests goes to playwright.spec.js.

This will help isplaywrightready and future jest integration.
